+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Added clearer in-line privacy notices when a 3rd party service is being used
 - Added `validate` mode to `sts-tts-provider-yaml` script to identify missing / additional / duplicate speakers in .yaml voice configuration, as well as validate fields with `--strict` flag
 - Added `sts-copy-to-clipboard` utility to copy file contents to clipboard
-- Added `sts-generate-voice-casting-prompt-file` utility to aid in casting voice configs
+- Added `sts-generate-character-notes-prompt` utility to aid in casting voice configs
 
 ## [1.1.0] - 2025-05-26